+++ src/Logic.js
@@ -8,43 +8,43 @@
  */
 Scrawler.Logic = function(args, callback, callbackArgs = []){
 
-	const self = this;
+	const _this_ = this;
 
-	self.id = args.id;
-	self.el = args.el;
-	self.order = args.order || 0;
-	self.range = args.range || null; // Array(2) with From and To values. 
-	self.baseline = args.baseline||0;
-	self.callback = callback;
-	self.callbackArgs = callbackArgs;
-	self.nodelist = document.querySelectorAll(args.el);
-	self.units = [];
-	for (var i = 0; i < self.nodelist.length; i++) {
-		self.units[i] = new Scrawler.Unit({
-			el: self.nodelist[i],
-			baseline: Common.calcBaseline(self.baseline, self.nodelist[i]),
+	_this_.id = args.id;
+	_this_.el = args.el;
+	_this_.order = args.order || 0;
+	_this_.range = args.range || null; // Array(2) with From and To values. 
+	_this_.baseline = args.baseline||0;
+	_this_.callback = callback;
+	_this_.callbackArgs = callbackArgs;
+	_this_.nodelist = document.querySelectorAll(args.el);
+	_this_.units = [];
+	for (var i = 0; i < _this_.nodelist.length; i++) {
+		_this_.units[i] = new Scrawler.Unit({
+			el: _this_.nodelist[i],
+			baseline: Common.calcBaseline(_this_.baseline, _this_.nodelist[i]),
 			progress: new Scrawler.Position()
 		});
 	}
 
-	self._range_unit_;
+	_this_._range_unit_;
 
-	if (self.range) {
-		if (typeof self.range[1] === 'string') {
-			if (self.range[1].indexOf('%') !== -1) {
+	if (_this_.range) {
+		if (typeof _this_.range[1] === 'string') {
+			if (_this_.range[1].indexOf('%') !== -1) {
 				// percent
-				self.range[0] = parseFloat(self.range[0].replace('%','')) / 100;
-				self.range[1] = parseFloat(self.range[1].replace('%','')) / 100;
-				self._range_unit_ = 'f';
-			} else if (self.range[1].indexOf('f') !== -1) {
-				self.range[0] = parseFloat(self.range[0].replace('f',''));
-				self.range[1] = parseFloat(self.range[1].replace('f',''));
-				self._range_unit_ = 'f';
+				_this_.range[0] = parseFloat(_this_.range[0].replace('%','')) / 100;
+				_this_.range[1] = parseFloat(_this_.range[1].replace('%','')) / 100;
+				_this_._range_unit_ = 'f';
+			} else if (_this_.range[1].indexOf('f') !== -1) {
+				_this_.range[0] = parseFloat(_this_.range[0].replace('f',''));
+				_this_.range[1] = parseFloat(_this_.range[1].replace('f',''));
+				_this_._range_unit_ = 'f';
 			} else {
-				self._range_unit_ = 'px';
+				_this_._range_unit_ = 'px';
 			}
 		} else {
-			self._range_unit_ = 'px';
+			_this_._range_unit_ = 'px';
 		}
 	}
 };